app/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
feedparser==5.2.1
Flask==1.0.2
Flask-Moment==0.6.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
python-dateutil==2.7.5
pytz==2013b0
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
